fix: split prompt workspace form under max-lines ratchet
Keep the New Workspace From Prompt form under the 500-line static check by moving attachment picking into a dedicated part file.
