+++ mobile/lib/src/features/workbench/presentation/create_workspace_screen.dart
@@ -25,6 +25,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 part 'create_workspace_manual.dart';
 part 'create_workspace_prompt.dart';
+part 'create_workspace_prompt_attachments.dart';
 
 class CreateWorkspaceScreen extends ConsumerStatefulWidget {
   const CreateWorkspaceScreen({
